requirements.txt
# Core LangChain packages (v0.3+)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
langchain-anthropic>=0.2.0

# PDF and document processing
pypdf>=3.17.0

# Vector stores and embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Environment and configuration
python-dotenv>=1.0.0
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# Data analysis and visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
scipy>=1.10.0

# Advanced workflows
langgraph>=0.2.0

# Evaluation framework (optional)
ragas>=0.1.0

# Jupyter support (if running in notebooks)
ipython>=8.0.0
jupyter>=1.0.0

# Excel support
openpyxl>=3.1.0